+++ apps/werkbank_werkbank/lib/main.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:werkbank/src/_internal/src/localizations/localizations.dart';
 import 'package:werkbank/werkbank.dart';
 import 'package:werkbank_werkbank/nms/nms_logo.dart';
-import 'package:werkbank_werkbank/sections.dart';
+import 'package:werkbank_werkbank/root.dart';
 
 void main() {
   runApp(const WerkbankWerkbank());
@@ -116,7 +116,7 @@ class WerkbankWerkbank extends StatelessWidget {
       logo: const NmsLogo(),
       lastUpdated: lastUpdated,
       appConfig: AppConfig.material(),
-      sections: sections,
+      root: root,
       addonConfig: addons,
     );
   }