+++ packages/werkbank/CHANGELOG.md
@@ -1,6 +1,13 @@
-# 0.15.2
+
+# 0.16.0
 - Add `c.background.colorBuilder(...)` and `c.background.widgetBuilder(...)` to allow easier use of theme colors in backgrounds.
 - Fix that pointer could interact with use case while trying to change constraints.
+- Make small changes to the tree UI to make it easier to overview the items and find the selected item.
+- Rename `WerkbankSections` to `WerkbankRoot`. **(BREAKING CHANGE)**
+  - This also means that the `sections` parameter of `WerkbankApp` has been renamed to `root`.
+- Rename `UseCaseControlSection` to `ConfigureControlSection` and `InfoControlSection` to `InspectControlSection`. **(BREAKING CHANGE)**
+  - This change should only affect custom addon authors.
+  - The tabs had been renamed from "USE CASE" to "CONFIGURE" and from "INFO" to "INSPECT" some time ago. However these classes were missed.
 
 # 0.15.1
 - Make some improvements to use case descriptions